Nft_Multiplier_Python_AI/requirements.txt
supabase>=2.3.1
requests>=2.31.0
google-generativeai>=0.3.2